+++ plone/app/blocks/tests/test_utils.py
@@ -47,3 +47,8 @@ class TestUtilsFunctional(unittest.TestCase):
     def test_resolve_resource_with_query(self):
         layout = resolveResource('/++sitelayout++testlayout1/site.html?ajax_load=1')
         self.assertIn(u'Layout panel 1', layout)
+
+    def test_resolve_resource_with_html_content(self):
+        # make sure we can handle content containing other ++ urls
+        layout = resolveResource('/?content=%3Cp%3ELorem%20ipsum%20dolor%20sit%20amet,%20consectetuer%20adipiscing%20elit.%20Sed%20posuere%20interdum%20sem.%20Quisque%20ligula%20eros%20ullamcorper%20quis,%20lacinia%20quis%20facilisis%20sed%20sapien.%3C/p%3E%0A%3Cscript%20type=%22application/javascript%22%20id=%22protect-script%22%20src=%22../../../++resource++protect.js')
+        self.assertIn(u'Plone Open Source CMS', layout)